+++ spyderlib/utils/introspection/__init__.py
@@ -8,5 +8,5 @@
 Introspection utilities used by Spyder
 """
 from . import module_completion
-from .plugin_manager import PluginManager
+from .manager import IntrospectionManager
