+++ spyderlib/utils/introspection/manager.py
